Add sniff that detects a quirk of microtime
By default, `microtime` has a string as its return value ("msec sec"), which
is unexpected and cannot be naively cast to float, making it error-prone. This
sniff disallows calls to `microtime()` without the `$get_as_float` argument
being explicitly set. It is still possible to set this argument to `false`,
but in that case you have probably thought about this.