Allow a small set of slevomat/coding-standard 7.0.x releases
We do not use a simple '^7' version constraint as the slevomat
helpers have become internal as of version 7. This means that
that they can break BC for the helpers (or just remove them
entirely) within the 7.x-range, which would break our sniffs.
The currently-released versions (7.0.0-7.0.4) still seem
compatible, so for now can allow this limited range.

See https://github.com/Moxio/php-codesniffer-sniffs/pull/10#issuecomment-822543540 for details.

Closes #10